+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/content/DownloadState.kt
@@ -57,6 +57,14 @@ data class DownloadState(
     val filePath: String
         get() = directoryPath + File.separatorChar + fileName
 
+    val isPdf: Boolean
+        get() =
+            if (contentType == null) {
+                fileName?.let { File(it).extension } == "pdf"
+            } else {
+                contentType == "application/pdf"
+            }
+
     /**
      * Status that represents every state that a download can be in.
      */